Created shared error component for form pages
