+++ client/app/redux/actions/index.js
@@ -1,0 +1,5 @@
+export * from "./alert";
+export * from "./app";
+export * from "./auth";
+export * from "./character";
+export * from "./game";
